Update package structure for v2 only
