+++ .gitignore
@@ -28,8 +28,8 @@ _testmain.go
 
 # Ignore test code coverage
 coverage.txt
-v2/vcr.coverprofile
-v2/cassette.coverprofile
-v2/recorder.coverprofile
+vcr.coverprofile
+cassette.coverprofile
+recorder.coverprofile
 
 .idea/
