+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2015-2016 Marin Atanasov Nikolov <dnaeon@gmail.com>
+Copyright (c) 2015-2022 Marin Atanasov Nikolov <dnaeon@gmail.com>
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
