+++ examples/fixtures/golang-org.yaml
